imporved the code of the projects api and update profile and made them more readable
