+++ controllersAPI/projects/projects.ts
@@ -1,6 +1,10 @@
 import project from "../../Model/project";
+import jwt from "jsonwebtoken";
 
-export const addTaskToProject = (req, res, data) => {
+let SECRET = process.env.REACT_APP_JWT_SECRET;
+
+export const addTaskToProject = (req, res) => {
+  let data = req.body;
   project.findOne({ _id: data.projectId }, async (err, result) => {
     let task = {
       title: data.TaskName,
@@ -26,3 +30,49 @@ export const addTaskToProject = (req, res, data) => {
     res.send(data);
   });
 };
+
+
+
+export const getUserProjects = (req,res)=>{
+  let cookies = req.cookies;
+
+    jwt.verify(cookies.auth, SECRET, (err, decoded) => {
+      if (!err && decoded) {
+        project.find({}, (err, docs) => {
+          let usersProjects = docs.filter((i) => i.team.includes(decoded.name));
+          res.send(usersProjects);
+        });
+      }
+    });
+}
+
+
+
+export const createNewProject =(req,res) =>{
+  let data = req.body;
+  let cookies = req.cookies;
+
+  jwt.verify(cookies.auth, SECRET, (err, decoded) => {
+    if (!err && decoded) {
+      project.create({
+        projectName: data.projectName,
+        projectDepartment: data.projectDepartment,
+        projectManager: data.projectManager,
+        team: [decoded.name],
+        toDo: [],
+        doing: [],
+        done: [],
+      });
+    } else {
+      res.send("Sign In");
+    }
+  });
+}
+
+
+
+export const deleteProject = async(req,res) =>{
+  let projectId = req.query.projectId;
+  await project.deleteOne({ _id: projectId });
+  res.send("project deleted");
+}
